+++ src/adaptors/api/yuque/yuqueApiAdaptor.ts
@@ -299,7 +299,7 @@ class YuqueApiAdaptor extends BaseBlogApi {
 
     // 使用兼容的fetch调用并返回统一的JSON数据
     const body = ObjectUtil.isEmptyObject(params) ? "" : JSON.stringify(params)
-    const resJson = await this.apiProxyFetch(apiUrl, [headers], body, method, contentType)
+    const resJson = await this.apiFetch(apiUrl, [headers], body, method, contentType)
     this.logger.debug("向语雀请求数据，resJson =>", resJson)
 
     if (resJson?.status === 401) {