+++ src/adaptors/web/csdn/csdnWebAdaptor.ts
@@ -25,11 +25,11 @@
 
 import { BaseWebApi } from "~/src/adaptors/web/base/baseWebApi.ts"
 import CsdnUtils from "~/src/adaptors/web/csdn/csdnUtils.ts"
-import { BlogConfig, CategoryInfo, PageTypeEnum, Post, UserBlog } from "zhi-blog-api"
+import {BlogConfig, CategoryInfo, MediaObject, PageTypeEnum, Post, UserBlog} from "zhi-blog-api"
 import { JsonUtil } from "zhi-common"
 import WebUtils from "~/src/adaptors/web/base/webUtils.ts"
 import _ from "lodash-es"
-import FormDataUtils from "~/src/utils/FormDataUtils.ts";
+import FormDataUtils from "~/src/utils/FormDataUtils.ts"
 
 /**
  * CSDN网页授权适配器
@@ -303,8 +303,18 @@ class CsdnWebAdaptor extends BaseWebApi {
     return flag
   }
 
-  public async uploadFile(file: File | Blob, filename?: string): Promise<any> {
+  public async getPreviewUrl(postid: string): Promise<string> {
+    const token = this.cfg.password
+    const userid = WebUtils.readCookie("UserName", token)
+    const previewUrl = this.cfg.previewUrl.replace(/\[userid\]/g, userid).replace(/\[postid\]/g, postid)
+    return previewUrl
+    // return StrUtil.pathJoin(this.cfg.home ?? "", previewUrl)
+  }
+
+  public async uploadFile(mediaObject: MediaObject): Promise<any> {
     const { FormData, Blob } = FormDataUtils.getFormData(this.appInstance)
+    const file = new Blob([mediaObject.bits], { type: mediaObject.type })
+    const filename = mediaObject.name
 
     this.logger.debug(`csdn start uploadFile ${filename}=>`, file)
     if (file instanceof Blob) {
@@ -325,13 +335,8 @@ class CsdnWebAdaptor extends BaseWebApi {
       formData.append("file", file)
 
       this.logger.debug("csdn image upload strat...")
-      const response = await fetch(uploadUrl, {
-        method: "POST",
-        body: formData,
-      })
-      const resText = await response.text()
-      this.logger.debug("csdn image upload success, resText=>", resText)
-      const resJson = JsonUtil.safeParse<any>(resText, {} as any)
+      const headers = {}
+      const resJson = await this.csdnFormFetch(uploadUrl, formData, headers)
       if (resJson.code !== 200) {
         throw new Error("CSDN图片上传失败 =>" + filename)
       }
@@ -340,10 +345,6 @@ class CsdnWebAdaptor extends BaseWebApi {
         object_key: resJson.data.targetObjectKeyy,
         url: resJson.data.imageUrl,
       }
-
-      // 其他方式，待研究。上面的仅PC客户端可用
-      // var res = await this.csdnFetch(uploadUrl, [], params, "POST", "multipart/form-data")
-      // this.logger.debug("csdn image upload success, res=>", res)
     }
 
     return {}
@@ -403,6 +404,7 @@ class CsdnWebAdaptor extends BaseWebApi {
     reqHeaderMap.set("x-ca-nonce", xCaNonce)
     reqHeaderMap.set("x-ca-signature", xCaSignature)
     reqHeaderMap.set("x-ca-signature-headers", "x-ca-key,x-ca-nonce")
+    reqHeaderMap.set("Cookie", this.cfg.password)
 
     const mergedHeaders = {
       ...Object.fromEntries(reqHeaderMap),
@@ -419,16 +421,36 @@ class CsdnWebAdaptor extends BaseWebApi {
     // 发送请求并返回响应
     this.logger.debug("csdn url =>", url)
     this.logger.debug("csdn requestOptions =>", requestOptions)
-    const res = await this.webProxyFetch(url, [mergedHeaders], params, method, contentType)
+    const res = await this.webFetch(url, [mergedHeaders], params, method, contentType, true, "base64")
     return res
   }
 
-  public async getPreviewUrl(postid: string): Promise<string> {
-    const token = this.cfg.password
-    const userid = WebUtils.readCookie("UserName", token)
-    const previewUrl = this.cfg.previewUrl.replace(/\[userid\]/g, userid).replace(/\[postid\]/g, postid)
-    return previewUrl
-    // return StrUtil.pathJoin(this.cfg.home ?? "", previewUrl)
+  /**
+   * 向 CSDN 发送表单数据
+   *
+   * @param url 请求地址
+   * @param formData 表单数据，默认为undefined，支持 ReadableStream、Blob | BufferSource | FormData | URLSearchParams | string。这里只需要 FormData
+   * @param headers 请求头
+   */
+  private async csdnFormFetch(url: string, formData: FormData, headers: Record<any, any> = {}) {
+    const apiUrl = url
+
+    const options: RequestInit = {
+      method: "POST",
+      headers: headers,
+      body: formData,
+    }
+
+    this.logger.debug("向 CSDN 发送表单数据，apiUrl =>", apiUrl)
+    this.logger.debug("向 CSDN 发送表单数据，options =>", options)
+
+    const resJson = await this.webFormFetch(apiUrl, [headers], formData, false)
+    if (resJson.error) {
+      throw new Error("CSDN 表单提交错误。详细错误 =>" + resJson.error)
+    }
+    this.logger.debug("向 CSDN 发送表单数据，resJson =>", resJson)
+
+    return resJson
   }
 }