+++ src/adaptors/web/csdn/useCsdnWeb.ts
@@ -86,6 +86,11 @@ const useCsdnWeb = async (key?: string, newCfg?: CsdnConfig) => {
   // picbed service
   cfg.picgoPicbedSupported = false
   cfg.bundledPicbedSupported = true
+  // 退出登录
+  cfg.logoutUrl = "https://passport.csdn.net/account/logout"
+  // picbed service
+  cfg.picgoPicbedSupported = false
+  cfg.bundledPicbedSupported = true
 
   const webApi = new CsdnWebAdaptor(appInstance, cfg)
   return {